backend/requirements.txt
cachetools
certifi
chardet
click
Flask
google-auth
google-auth-oauthlib
idna
itsdangerous
Jinja2
MarkupSafe
oauthlib
pkg-resources
pyasn1
pyasn1-modules
requests
requests-oauthlib
rsa
six
urllib3
Werkzeug
